ci: split UI tests for the input page in 3 groups
